requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==3.2.5
django-bootstrap4==23.1
django-ckeditor==6.2.0
django-crispy-forms==2.0
django-js-asset==2.0.0
paytmchecksum==1.7.0
Pillow==9.5.0
pycryptodome==3.16.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
APScheduler==3.10.1
python-dateutil==2.8.2
